HotFix1 - Arreglado migraciones para que trabajen correctamente con InnoDB. Se arreglan los tipos de las variables a los correspondientes que se utilizarán.
 - Modificado todos los controladores para ordernarse por ID descendientemente y creadas formas para mostrar los datos y mensajes de error personalizados, así como tambien la exclusión de la categoria 0 (general) y asignación de ésta al borrar. No se permite crear correo ni cuenta mientras no se configure los datos de empresa.
 - Seteado el timezone para trabajar con la función now() de php, y guardar la fecha de modificación del producto, y la fecha de creación del cupón.
 - Se modifican validaciones correspondientes y se añaden atributos correspondientes a modelo de cupón.
 - Modificado crear.js para trabajar con los distintos select y datetimepicker de las vistas.
 - Modificada función index para que imprima mensaje de error en esquina cuando sea else (no es request ajax).
 - initInputMask.js modificado para que cuando se vacíe el campo de rut, retorne vacío con el placeholder y no "-undefined".
 - Modificado el Select2 para que trabaje con modals y quitado barra de búsqueda de elementos sin muchos option.
 - Modificada todas las vistas para trabajar con los select2, y poder desplegar correctamente los datos en tablas.
 - Se remueve ID y se trabaja con Index+1++.
 - Añadido un mensaje de error para ciertos errores (no configurada empresa al momento de crear correo y cuenta) en mensaje.blade.php.
 - Modificado el aside.blade, para que asigne la clase active si se encuentra en la ruta correspondiente.
 - Agregado al layout un yield de breadcrumbs, el cuál se agregó a cada vista.

+++ database/migrations/2019_08_02_020713_creartabla_detalle_venta.php
@@ -18,8 +18,8 @@ class CreartablaDetalleVenta extends Migration
             $table->integer('cantidad');
             $table->integer('subtotal');
             $table->integer('precio_venta_unitario');
-            $table->unsignedInteger('producto_id');
-            $table->unsignedInteger('venta_id');
+            $table->unsignedBigInteger('producto_id');
+            $table->unsignedBigInteger('venta_id');
             $table->foreign('producto_id')->references('id')->on('producto')->onDelete('restrict')->onUpdate('restrict');
             $table->foreign('venta_id')->references('id')->on('venta')->onDelete('restrict')->onUpdate('restrict');
             $table->charset = 'utf8mb4';
